requirements.txt
beautifulsoup4
requests
botogram2
jellyfish
